Cylink/app/requirements.txt
kivy==2.1.0
pyjnius
kivy_garden.graph
